requirements.txt
Django==1.7
ajaxuploader==0.3.8
argparse==1.2.1
astroid==1.2.1
coverage==3.7.1
django-admin-bootstrapped==2.3.5
django-ajax-upload-widget==0.5.5
django-appconf==0.6
django-bootstrap3==4.11.0
django-compressor==1.4
django-jenkins==0.16.3
django-http-proxy==0.4
ipython==2.3.0
logilab-common==0.62.1
pep8==1.5.7
pylint==1.3.1
six==1.8.0
wsgiref==0.1.2